requirements.txt
streamlit
pandas
pillow
plotly
neo4j